feat: Updating Readme
Creating a Readme based on a Readme template. All information about the project was added to the Readme, getting information from the initial repository made for Android and the commits from the current repository.

This documentation needs the complementation of someone who understands Flutter, this PR is just the basis of the documentation!

A preview can be seen at:
https://gist.github.com/csorlandi/1324df24b731d0802e6d5c86d1de5c8f
(just without the images)